EOS-20944-SASportHealthView: Added SASInterface class and common error class.
