+++ low-level/framework/utils/sysfs_interface.py
@@ -141,133 +141,3 @@ class SysFS(Utility):
         except OSError as os_err:
             return os_err.errno
         return self.nw_phy_link_state[carrier_indicator]
-
-    def get_sas_host_list(self):
-        """
-            Return SAS host list from /sys/class/sas_host directory.
-            eg: ['host1', 'host2']
-        """
-        sas_hosts = []
-        err_str = ""
-        err_no = 0
-        try:
-            host_dirpath = self.get_sys_dir_path('sas_host')
-            sas_hosts = os.listdir(host_dirpath)
-            sas_hosts = [host for host in sas_hosts
-                         if 'host' in host]
-        except Exception as err:
-            sas_hosts = []
-            err_no, err_str = err.args
-        return sas_hosts, err_str, err_no
-
-    def get_sas_port_list(self, host=""):
-        """
-            Return a list of SAS ports.
-            searches for ports from given host in /sys/class/sas_port
-            directory if directory exists else return an empty list.
-            eg: ['port-1:0', 'port-1:1', 'port-1:2', 'port-1:3']
-        """
-        sas_ports = []
-        err_str = ""
-        err_no = 0
-        try:
-            host_id = host.replace('host', '')
-            port_dirpath = self.get_sys_dir_path('sas_port')
-            sas_ports = os.listdir(port_dirpath)
-            sas_ports = [port for port in sas_ports
-                         if f'port-{host_id}' in port]
-        except Exception as err:
-            sas_ports = []
-            err_no, err_str = err.args
-        return sas_ports, err_str, err_no
-
-    def get_sas_port_data(self, port_name):
-        """
-            Return SAS port data/info of given port.
-            eg: Given input port-1:0 returns
-                { "port_id":"sas_port-0",
-                  "state":"running",
-                  "sas_address":"0x500c0fff0a98b000"
-                }
-        """
-        port_data = {}
-        err_str = ""
-        err_no = 0
-        try:
-            host_id, port_id = port_name.replace('port-', '').split(':')
-            port_data_path = os.path.join(
-                             self.get_sys_dir_path('sas_end_device'),
-                             f'end_device-{host_id}:{port_id}',
-                             f'device/target{host_id}:0:{port_id}',
-                             f'{host_id}:0:{port_id}:0')
-            with open(os.path.join(port_data_path, 'state')) as cFile:
-                state = cFile.read().strip()
-
-            with open(os.path.join(port_data_path, 'sas_address')) as cFile:
-                sas_addr = cFile.read().strip()
-
-            port_data = {
-                    "port_id": f'sas_port-{port_id}',
-                    "state": state,
-                    "sas_address": sas_addr
-                }
-        except Exception as err:
-            port_data = {}
-            err_no, err_str = err.args
-        return port_data, err_str, err_no
-
-    def get_phy_list_for_port(self, port):
-        """
-            Return list of SAS phys under given port.
-            eg: for given port-1:0 it might return
-                [ 'phy-1:0', 'phy-1:1', 'phy-1:2', 'phy-1:3']
-                returns an empty list if port-1:0 does not exist.
-        """
-        phys = []
-        err_str = ""
-        err_no = 0
-        try:
-            phy_list_dir = os.path.join(
-                            self.get_sys_dir_path('sas_port'),
-                            port, 'device')
-            phys = os.listdir(phy_list_dir)
-            phys = [phy for phy in phys if 'phy' in phy]
-            phys.sort(key=lambda phy: int(phy.split(':')[1]))
-        except Exception as err:
-            phys = []
-            err_no, err_str = err.args
-        return phys, err_str, err_no
-
-    def get_sas_phy_data(self, phy):
-        """
-            Return SAS phy information.
-            eg: for phy `phy-1:8` it might return
-                {
-                    "phy_id":"phy-1:8",
-                    "state":"enabled",
-                    "negotiated_linkrate":"12.0 Gbit"
-                }
-        """
-        phy_data = {}
-        err_str = ""
-        err_no = 0
-        try:
-            phy_data_dir = os.path.join(
-                            self.get_sys_dir_path('sas_phy'), phy)
-
-            with open(os.path.join(phy_data_dir, 'enable')) as cFile:
-                state = cFile.read().strip()
-                state = "enabled" if state == '1' else "disabled"
-
-            with open(os.path.join(phy_data_dir, 'negotiated_linkrate')) as cFile:
-                n_link_rate = cFile.read().strip()
-
-            phy_data = {
-                "phy_id": phy,
-                "state": state,
-                "negotiated_linkrate": n_link_rate
-            }
-        except Exception as err:
-            phy_data = {}
-            err_no, err_str = err.args
-        return phy_data, err_str, err_no
